spotify-producer/requirements.txt
requests==2.21.0
kafka-python==1.4.5
